update remaining examples

+++ examples/volumedelete.rs
@@ -1,8 +1,8 @@
 use shiplift::Docker;
 use std::env;
-use tokio::prelude::Future;
 
-fn main() {
+#[tokio::main]
+async fn main() {
     let docker = Docker::new();
     let volumes = docker.volumes();
 
@@ -10,11 +10,7 @@ fn main() {
         .nth(1)
         .expect("You need to specify an volume name");
 
-    let fut = volumes
-        .get(&volume_name)
-        .delete()
-        .map(|info| println!("{:?}", info))
-        .map_err(|e| eprintln!("Error: {}", e));
-
-    tokio::run(fut);
+    if let Err(e) = volumes.get(&volume_name).delete().await {
+        eprintln!("Error: {}", e)
+    }
 }
